requirements.txt
streamlit
opencv-python
numpy
requests
easyocr
torch
torchvision
Pillow
python-dotenv